Install raspi-config if missing
When run on something other than raspbian, raspi-config likely won't be
installed. This installs the latest version of the raspi-config DEB from
the Raspbian repo.

The reasons for installing the single package rather than adding the
raspbian repo is due to the potential for disaster that introduces. While
you can use pinning, if the pin config gets removed, you will likely have
a broken system after doing an apt-get upgrade.